k8s: Use ParseService when comparing two services
Previously, the EqualV1Services() function had its own k8s.Service
constructor which was not in sync with ParseService() and did not
consider some service fields.

The consequence of this was that changing a type or ports of a service
did not trigger the service update, and thus the service change was
not reflected in the SVC BPF maps.

Use ParseService() instead of the custom constructor to prevent any
discrepancies in the future.

Signed-off-by: Martynas Pumputis <m@lambda.lt>