k8s: Add support for LoadBalancer services
This commit adds a support for a service of the LoadBalancer type when
running w/o kube-proxy.

Such service has an IP addr assigned by an external loadbalancer (e.g.
MetalLB) which might be used to access a service from outside. The IP
addr is assigned not immediately. Therefore, the provisioning of the
service usually takes place after receiving a second update of the service
which contains the IP addr stored in the Service.Status.LoadBalancer.Ingress
field (array).

The behavior of the service is identical to a service of the ExternalIP
"type", therefore no changes are required in the BPF datapath.

Signed-off-by: Martynas Pumputis <m@lambda.lt>