node: Use non-local-node identity when enabled
Assign the new non-local-node identity to all remote hosts if the option is
enabled. The option is disabled by default as it has the potential to change
existing policy rules where users are relying on the behavior that allowing
from the host will allow traffic from all hosts in the cluster.

Signed-off-by: Thomas Graf <thomas@cilium.io>